requirements.txt
django==3.2.9
djangorestframework==3.13.1
markdown==3.3.7
django-filter==21.1
sphinx==5.0.0
whitenoise==6.1.0
psycopg2-binary==2.9.6